# Avoid frequent polling in OSMInputFile