This PR tries to avoid the locking overhead caused by polling the LinkedBlockingQueue for each element by draining up to 1k available elements into an ArrayDeque.