@karussell 2020-11-25
Should this be faster and/or did you benchmark the code?