@otbutz 2020-11-25
I tested it with an Austria pbf (596MB)

Master:
```
2020-11-25 21:40:07.645 [main] INFO  com.graphhopper.reader.osm.OSMReader - 100 000 (preprocess), osmWayMap:467 156 totalMB:5120, usedMB:1553
2020-11-25 21:40:07.914 [main] INFO  com.graphhopper.reader.osm.OSMReader - creating graph. Found nodes (pillar+tower):13 271 268, totalMB:5120, usedMB:1827
2020-11-25 21:40:33.973 [main] INFO  com.graphhopper.reader.osm.OSMReader - 69 914 450, now parsing ways
2020-11-25 21:40:58.963 [main] INFO  com.graphhopper.reader.osm.OSMReader - 77 497 029, now parsing relations
2020-11-25 21:40:59.179 [main] INFO  com.graphhopper.reader.osm.OSMReader - finished way processing. nodes: 1979024, osmIdMap.size:13291849, osmIdMap:168MB, nodeFlagsMap.size:20581, relFlagsMap.size:619224, zeroCounter:20298 totalMB:5120, usedMB:3216
2020-11-25 21:40:59.179 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:23s, pass2:51s, total:74s
```

PR:
```
2020-11-25 21:34:16.921 [main] INFO  com.graphhopper.reader.osm.OSMReader - 100 000 (preprocess), osmWayMap:467 156 totalMB:5120, usedMB:1638
2020-11-25 21:34:17.253 [main] INFO  com.graphhopper.reader.osm.OSMReader - creating graph. Found nodes (pillar+tower):13 271 268, totalMB:5120, usedMB:1916
2020-11-25 21:34:28.660 [main] INFO  com.graphhopper.reader.osm.OSMReader - 69 914 450, now parsing ways
2020-11-25 21:34:48.808 [main] INFO  com.graphhopper.reader.osm.OSMReader - 77 497 029, now parsing relations
2020-11-25 21:34:49.081 [main] INFO  com.graphhopper.reader.osm.OSMReader - finished way processing. nodes: 1979024, osmIdMap.size:13291849, osmIdMap:168MB, nodeFlagsMap.size:20581, relFlagsMap.size:619224, zeroCounter:20298 totalMB:5120, usedMB:3023
2020-11-25 21:34:49.082 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:19s, pass2:31s, total:50s
```

It's a little bit faster :grin: