@karussell 2020-11-25
Impressive. I can reproduce this approx. 30% speed up :+1: